+++ src/__tests__/counter.unit.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, fireEvent } from '@testing-library/react';
-import Counter from './Counter';
+import Counter from '../components/Counter';
 
 describe('Counter Component', () => {
   it('should render the initial value provided in props', () => {
